
\name{basicplots}
\title{Basic benchmark experiment plots.}
\description{Basic plots for exploratory analysis of benchmark
  experiment.}
\details{
  Basic plots for the exploratory analysis of benchmark experiments. For
  investigations on \emph{algorithms} on one \emph{dataset} plots are
  implemented with the traditional graphics system. For investigations
  on \emph{algorithms} on more than one \emph{data set} lattice plots
  are implemented. The naming scheme of the plots indicate the the
  respective graphic systems and therefore the capabilities of the
  plots.
}

